Introduce IWarhead and move Warhead to Mods.Common
