+++ OpenRA.Game/OpenRA.Game.csproj
@@ -100,7 +100,6 @@
     <Compile Include="FileSystem\BagFile.cs" />
     <Compile Include="Map\MapPlayers.cs" />
     <Compile Include="MPos.cs" />
-    <Compile Include="GameRules\Warhead.cs" />
     <Compile Include="Graphics\QuadRenderer.cs" />
     <Compile Include="Download.cs" />
     <Compile Include="Effects\AsyncAction.cs" />
@@ -235,7 +234,6 @@
     <Compile Include="Graphics\SelectionBarsRenderable.cs" />
     <Compile Include="Graphics\TargetLineRenderable.cs" />
     <Compile Include="Graphics\UISpriteRenderable.cs" />
-    <Compile Include="GameRules\DamageWarhead.cs" />
     <Compile Include="Graphics\SoftwareCursor.cs" />
     <Compile Include="Graphics\HardwareCursor.cs" />
     <Compile Include="Support\PerfItem.cs" />
