+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -539,11 +539,13 @@
     <Compile Include="UtilityCommands\UpgradeRules.cs" />
     <Compile Include="Warheads\CreateEffectWarhead.cs" />
     <Compile Include="Warheads\CreateResourceWarhead.cs" />
+    <Compile Include="Warheads\DamageWarhead.cs" />
     <Compile Include="Warheads\DestroyResourceWarhead.cs" />
     <Compile Include="Warheads\GrantUpgradeWarhead.cs" />
     <Compile Include="Warheads\HealthPercentageDamageWarhead.cs" />
     <Compile Include="Warheads\LeaveSmudgeWarhead.cs" />
     <Compile Include="Warheads\SpreadDamageWarhead.cs" />
+    <Compile Include="Warheads\Warhead.cs" />
     <Compile Include="Widgets\ActorPreviewWidget.cs" />
     <Compile Include="Widgets\ColorMixerWidget.cs" />
     <Compile Include="Widgets\ColorPreviewManagerWidget.cs" />
